+++ src/reader/async_tokio.rs
@@ -425,6 +425,7 @@ mod test {
         read_event_into_async,
         read_until_close_async,
         TokioAdapter,
+        1,
         &mut Vec::new(),
         async,
         await