+++ src/reader/mod.rs
@@ -1237,6 +1237,7 @@ mod test {
             $read_until_close:ident,
             // constructor of the XML source on which internal functions will be called
             $source:path,
+            $skip:literal,
             // constructor of the buffer to which read data will stored
             $buf:expr
             $(, $async:ident, $await:ident)?
@@ -1258,9 +1259,9 @@ mod test {
                     #[ignore = "start CDATA sequence fully checked outside of `read_bang_element`"]
                     $($async)? fn not_properly_start() {
                         let buf = $buf;
-                        let mut position = 1;
-                        let mut input = b"<![]]>other content".as_ref();
-                        //                ^= 1
+                        let mut position = 0;
+                        let mut input = &b"<![]]>other content"[$skip..];
+                        //                 ^= 0
 
                         match $source(&mut input).read_bang_element(buf, &mut position) $(.$await)? {
                             Err(Error::Syntax(cause)) => assert_eq!(cause, SyntaxError::UnclosedCData),
@@ -1278,8 +1279,8 @@ mod test {
                     $($async)? fn not_closed() {
                         let buf = $buf;
                         let mut position = 0;
-                        let mut input = b"<![CDATA[other content".as_ref();
-                        //                ^= 0                  ^= 22
+                        let mut input = &b"<![CDATA[other content"[$skip..];
+                        //                 ^= 0                  ^= 22
 
                         match $source(&mut input).read_bang_element(buf, &mut position) $(.$await)? {
                             Err(Error::Syntax(cause)) => assert_eq!(cause, SyntaxError::UnclosedCData),
@@ -1296,7 +1297,7 @@ mod test {
                     $($async)? fn empty() {
                         let buf = $buf;
                         let mut position = 0;
-                        let mut input = b"<![CDATA[]]>other content".as_ref();
+                        let mut input = &b"<![CDATA[]]>other content"[$skip..];
                         //                ^= 0        ^= 12
 
                         let (ty, bytes) = $source(&mut input)
@@ -1317,8 +1318,8 @@ mod test {
                     $($async)? fn with_content() {
                         let buf = $buf;
                         let mut position = 0;
-                        let mut input = b"<![CDATA[cdata]] ]>content]]>other content]]>".as_ref();
-                        //                ^= 0                         ^= 29
+                        let mut input = &b"<![CDATA[cdata]] ]>content]]>other content]]>"[$skip..];
+                        //                 ^= 0                         ^= 29
 
                         let (ty, bytes) = $source(&mut input)
                             .read_bang_element(buf, &mut position)
@@ -1356,9 +1357,9 @@ mod test {
                     #[ignore = "start comment sequence fully checked outside of `read_bang_element`"]
                     $($async)? fn not_properly_start() {
                         let buf = $buf;
-                        let mut position = 1;
-                        let mut input = b"<!- -->other content".as_ref();
-                        //                ^= 1
+                        let mut position = 0;
+                        let mut input = &b"<!- -->other content"[$skip..];
+                        //                  ^= 1
 
                         match $source(&mut input).read_bang_element(buf, &mut position) $(.$await)? {
                             Err(Error::Syntax(cause)) => assert_eq!(cause, SyntaxError::UnclosedComment),
@@ -1374,8 +1375,8 @@ mod test {
                     $($async)? fn not_properly_end() {
                         let buf = $buf;
                         let mut position = 0;
-                        let mut input = b"<!->other content".as_ref();
-                        //                ^= 0             ^= 17
+                        let mut input = &b"<!->other content"[$skip..];
+                        //                 ^= 0             ^= 17
 
                         match $source(&mut input).read_bang_element(buf, &mut position) $(.$await)? {
                             Err(Error::Syntax(cause)) => assert_eq!(cause, SyntaxError::UnclosedComment),
@@ -1391,8 +1392,8 @@ mod test {
                     $($async)? fn not_closed1() {
                         let buf = $buf;
                         let mut position = 0;
-                        let mut input = b"<!--other content".as_ref();
-                        //                ^= 0             ^= 17
+                        let mut input = &b"<!--other content"[$skip..];
+                        //                 ^= 0             ^= 17
 
                         match $source(&mut input).read_bang_element(buf, &mut position) $(.$await)? {
                             Err(Error::Syntax(cause)) => assert_eq!(cause, SyntaxError::UnclosedComment),
@@ -1408,8 +1409,8 @@ mod test {
                     $($async)? fn not_closed2() {
                         let buf = $buf;
                         let mut position = 0;
-                        let mut input = b"<!-->other content".as_ref();
-                        //                ^= 0              ^= 18
+                        let mut input = &b"<!-->other content"[$skip..];
+                        //                 ^= 0              ^= 18
 
                         match $source(&mut input).read_bang_element(buf, &mut position) $(.$await)? {
                             Err(Error::Syntax(cause)) => assert_eq!(cause, SyntaxError::UnclosedComment),
@@ -1425,8 +1426,8 @@ mod test {
                     $($async)? fn not_closed3() {
                         let buf = $buf;
                         let mut position = 0;
-                        let mut input = b"<!--->other content".as_ref();
-                        //                ^= 0               ^= 19
+                        let mut input = &b"<!--->other content"[$skip..];
+                        //                 ^= 0               ^= 19
 
                         match $source(&mut input).read_bang_element(buf, &mut position) $(.$await)? {
                             Err(Error::Syntax(cause)) => assert_eq!(cause, SyntaxError::UnclosedComment),
@@ -1442,8 +1443,8 @@ mod test {
                     $($async)? fn empty() {
                         let buf = $buf;
                         let mut position = 0;
-                        let mut input = b"<!---->other content".as_ref();
-                        //                ^= 0   ^= 7
+                        let mut input = &b"<!---->other content"[$skip..];
+                        //                 ^= 0   ^= 7
 
                         let (ty, bytes) = $source(&mut input)
                             .read_bang_element(buf, &mut position)
@@ -1460,8 +1461,8 @@ mod test {
                     $($async)? fn with_content() {
                         let buf = $buf;
                         let mut position = 0;
-                        let mut input = b"<!--->comment<--->other content".as_ref();
-                        //                ^= 0              ^= 18
+                        let mut input = &b"<!--->comment<--->other content"[$skip..];
+                        //                 ^= 0              ^= 18
 
                         let (ty, bytes) = $source(&mut input)
                             .read_bang_element(buf, &mut position)
@@ -1487,8 +1488,8 @@ mod test {
                         $($async)? fn not_properly_start() {
                             let buf = $buf;
                             let mut position = 0;
-                            let mut input = b"<!D other content".as_ref();
-                            //                ^= 0             ^= 17
+                            let mut input = &b"<!D other content"[$skip..];
+                            //                 ^= 0             ^= 17
 
                             match $source(&mut input).read_bang_element(buf, &mut position) $(.$await)? {
                                 Err(Error::Syntax(cause)) => assert_eq!(cause, SyntaxError::UnclosedDoctype),
@@ -1504,8 +1505,8 @@ mod test {
                         $($async)? fn without_space() {
                             let buf = $buf;
                             let mut position = 0;
-                            let mut input = b"<!DOCTYPEother content".as_ref();
-                            //                ^= 0                  ^= 22
+                            let mut input = &b"<!DOCTYPEother content"[$skip..];
+                            //                 ^= 0                  ^= 22
 
                             match $source(&mut input).read_bang_element(buf, &mut position) $(.$await)? {
                                 Err(Error::Syntax(cause)) => assert_eq!(cause, SyntaxError::UnclosedDoctype),
@@ -1521,8 +1522,8 @@ mod test {
                         $($async)? fn empty() {
                             let buf = $buf;
                             let mut position = 0;
-                            let mut input = b"<!DOCTYPE>other content".as_ref();
-                            //                ^= 0      ^= 10
+                            let mut input = &b"<!DOCTYPE>other content"[$skip..];
+                            //                 ^= 0      ^= 10
 
                             let (ty, bytes) = $source(&mut input)
                                 .read_bang_element(buf, &mut position)
@@ -1539,8 +1540,8 @@ mod test {
                         $($async)? fn not_closed() {
                             let buf = $buf;
                             let mut position = 0;
-                            let mut input = b"<!DOCTYPE other content".as_ref();
-                            //                ^= 0                   ^23
+                            let mut input = &b"<!DOCTYPE other content"[$skip..];
+                            //                 ^= 0                   ^23
 
                             match $source(&mut input).read_bang_element(buf, &mut position) $(.$await)? {
                                 Err(Error::Syntax(cause)) => assert_eq!(cause, SyntaxError::UnclosedDoctype),
@@ -1561,8 +1562,8 @@ mod test {
                         $($async)? fn not_properly_start() {
                             let buf = $buf;
                             let mut position = 0;
-                            let mut input = b"<!d other content".as_ref();
-                            //                ^= 0             ^= 17
+                            let mut input = &b"<!d other content"[$skip..];
+                            //                 ^= 0             ^= 17
 
                             match $source(&mut input).read_bang_element(buf, &mut position) $(.$await)? {
                                 Err(Error::Syntax(cause)) => assert_eq!(cause, SyntaxError::UnclosedDoctype),
@@ -1578,8 +1579,8 @@ mod test {
                         $($async)? fn without_space() {
                             let buf = $buf;
                             let mut position = 0;
-                            let mut input = b"<!doctypeother content".as_ref();
-                            //                ^= 0                  ^= 22
+                            let mut input = &b"<!doctypeother content"[$skip..];
+                            //                 ^= 0                  ^= 22
 
                             match $source(&mut input).read_bang_element(buf, &mut position) $(.$await)? {
                                 Err(Error::Syntax(cause)) => assert_eq!(cause, SyntaxError::UnclosedDoctype),
@@ -1595,8 +1596,8 @@ mod test {
                         $($async)? fn empty() {
                             let buf = $buf;
                             let mut position = 0;
-                            let mut input = b"<!doctype>other content".as_ref();
-                            //                ^= 0      ^= 10
+                            let mut input = &b"<!doctype>other content"[$skip..];
+                            //                 ^= 0      ^= 10
 
                             let (ty, bytes) = $source(&mut input)
                                 .read_bang_element(buf, &mut position)
@@ -1613,8 +1614,8 @@ mod test {
                         $($async)? fn not_closed() {
                             let buf = $buf;
                             let mut position = 0;
-                            let mut input = b"<!doctype other content".as_ref();
-                            //                ^= 0                   ^= 23
+                            let mut input = &b"<!doctype other content"[$skip..];
+                            //                 ^= 0                   ^= 23
 
                             match $source(&mut input).read_bang_element(buf, &mut position) $(.$await)? {
                                 Err(Error::Syntax(cause)) => assert_eq!(cause, SyntaxError::UnclosedDoctype),
@@ -1809,12 +1810,14 @@ mod test {
                 use pretty_assertions::assert_eq;
 
                 /// Checks that nothing was read from empty buffer
+                /// `<` read in peek_one that is called before read_with, that is why it in the input buffer
+                /// peek_one, however, does not increment position for simplicity of the code
                 #[$test]
                 $($async)? fn empty() {
                     let buf = $buf;
-                    let mut position = 1;
-                    let mut input = b"".as_ref();
-                    //                ^= 1
+                    let mut position = 0;
+                    let mut input = &b"<"[$skip..];
+                    //                  ^= 1
 
                     match $source(&mut input).read_with(ElementParser::default(), buf, &mut position) $(.$await)? {
                         Err(Error::Syntax(cause)) => assert_eq!(cause, SyntaxError::UnclosedTag),
@@ -1833,13 +1836,13 @@ mod test {
                     #[$test]
                     $($async)? fn empty_tag() {
                         let buf = $buf;
-                        let mut position = 1;
-                        let mut input = b">".as_ref();
-                        //                 ^= 2
+                        let mut position = 0;
+                        let mut input = &b"<>"[$skip..];
+                        //                   ^= 2
 
                         assert_eq!(
                             Bytes($source(&mut input).read_with(ElementParser::default(), buf, &mut position) $(.$await)? .unwrap()),
-                            Bytes(b"")
+                            Bytes(b"<")
                         );
                         assert_eq!(position, 2);
                     }
@@ -1847,13 +1850,13 @@ mod test {
                     #[$test]
                     $($async)? fn normal() {
                         let buf = $buf;
-                        let mut position = 1;
-                        let mut input = b"tag>".as_ref();
-                        //                    ^= 5
+                        let mut position = 0;
+                        let mut input = &b"<tag>"[$skip..];
+                        //                      ^= 5
 
                         assert_eq!(
                             Bytes($source(&mut input).read_with(ElementParser::default(), buf, &mut position) $(.$await)? .unwrap()),
-                            Bytes(b"tag")
+                            Bytes(b"<tag")
                         );
                         assert_eq!(position, 5);
                     }
@@ -1861,13 +1864,13 @@ mod test {
                     #[$test]
                     $($async)? fn empty_ns_empty_tag() {
                         let buf = $buf;
-                        let mut position = 1;
-                        let mut input = b":>".as_ref();
-                        //                  ^= 3
+                        let mut position = 0;
+                        let mut input = &b"<:>"[$skip..];
+                        //                    ^= 3
 
                         assert_eq!(
                             Bytes($source(&mut input).read_with(ElementParser::default(), buf, &mut position) $(.$await)? .unwrap()),
-                            Bytes(b":")
+                            Bytes(b"<:")
                         );
                         assert_eq!(position, 3);
                     }
@@ -1875,13 +1878,13 @@ mod test {
                     #[$test]
                     $($async)? fn empty_ns() {
                         let buf = $buf;
-                        let mut position = 1;
-                        let mut input = b":tag>".as_ref();
-                        //                     ^= 6
+                        let mut position = 0;
+                        let mut input = &b"<:tag>"[$skip..];
+                        //                       ^= 6
 
                         assert_eq!(
                             Bytes($source(&mut input).read_with(ElementParser::default(), buf, &mut position) $(.$await)? .unwrap()),
-                            Bytes(b":tag")
+                            Bytes(b"<:tag")
                         );
                         assert_eq!(position, 6);
                     }
@@ -1889,13 +1892,13 @@ mod test {
                     #[$test]
                     $($async)? fn with_attributes() {
                         let buf = $buf;
-                        let mut position = 1;
-                        let mut input = br#"tag  attr-1=">"  attr2  =  '>'  3attr>"#.as_ref();
-                        //                                                        ^= 39
+                        let mut position = 0;
+                        let mut input = &br#"<tag  attr-1=">"  attr2  =  '>'  3attr>"#[$skip..];
+                        //                                                          ^= 39
 
                         assert_eq!(
                             Bytes($source(&mut input).read_with(ElementParser::default(), buf, &mut position) $(.$await)? .unwrap()),
-                            Bytes(br#"tag  attr-1=">"  attr2  =  '>'  3attr"#)
+                            Bytes(br#"<tag  attr-1=">"  attr2  =  '>'  3attr"#)
                         );
                         assert_eq!(position, 39);
                     }
@@ -1908,13 +1911,13 @@ mod test {
                     #[$test]
                     $($async)? fn empty_tag() {
                         let buf = $buf;
-                        let mut position = 1;
-                        let mut input = b"/>".as_ref();
-                        //                  ^= 3
+                        let mut position = 0;
+                        let mut input = &b"</>"[$skip..];
+                        //                    ^= 3
 
                         assert_eq!(
                             Bytes($source(&mut input).read_with(ElementParser::default(), buf, &mut position) $(.$await)? .unwrap()),
-                            Bytes(b"/")
+                            Bytes(b"</")
                         );
                         assert_eq!(position, 3);
                     }
@@ -1922,13 +1925,13 @@ mod test {
                     #[$test]
                     $($async)? fn normal() {
                         let buf = $buf;
-                        let mut position = 1;
-                        let mut input = b"tag/>".as_ref();
-                        //                     ^= 6
+                        let mut position = 0;
+                        let mut input = &b"<tag/>"[$skip..];
+                        //                       ^= 6
 
                         assert_eq!(
                             Bytes($source(&mut input).read_with(ElementParser::default(), buf, &mut position) $(.$await)? .unwrap()),
-                            Bytes(b"tag/")
+                            Bytes(b"<tag/")
                         );
                         assert_eq!(position, 6);
                     }
@@ -1936,13 +1939,13 @@ mod test {
                     #[$test]
                     $($async)? fn empty_ns_empty_tag() {
                         let buf = $buf;
-                        let mut position = 1;
-                        let mut input = b":/>".as_ref();
-                        //                   ^= 4
+                        let mut position = 0;
+                        let mut input = &b"<:/>"[$skip..];
+                        //                     ^= 4
 
                         assert_eq!(
                             Bytes($source(&mut input).read_with(ElementParser::default(), buf, &mut position) $(.$await)? .unwrap()),
-                            Bytes(b":/")
+                            Bytes(b"<:/")
                         );
                         assert_eq!(position, 4);
                     }
@@ -1950,13 +1953,13 @@ mod test {
                     #[$test]
                     $($async)? fn empty_ns() {
                         let buf = $buf;
-                        let mut position = 1;
-                        let mut input = b":tag/>".as_ref();
-                        //                      ^= 7
+                        let mut position = 0;
+                        let mut input = &b"<:tag/>"[$skip..];
+                        //                        ^= 7
 
                         assert_eq!(
                             Bytes($source(&mut input).read_with(ElementParser::default(), buf, &mut position) $(.$await)? .unwrap()),
-                            Bytes(b":tag/")
+                            Bytes(b"<:tag/")
                         );
                         assert_eq!(position, 7);
                     }
@@ -1964,13 +1967,13 @@ mod test {
                     #[$test]
                     $($async)? fn with_attributes() {
                         let buf = $buf;
-                        let mut position = 1;
-                        let mut input = br#"tag  attr-1="/>"  attr2  =  '/>'  3attr/>"#.as_ref();
-                        //                                                           ^= 42
+                        let mut position = 0;
+                        let mut input = &br#"<tag  attr-1="/>"  attr2  =  '/>'  3attr/>"#[$skip..];
+                        //                                                             ^= 42
 
                         assert_eq!(
                             Bytes($source(&mut input).read_with(ElementParser::default(), buf, &mut position) $(.$await)? .unwrap()),
-                            Bytes(br#"tag  attr-1="/>"  attr2  =  '/>'  3attr/"#)
+                            Bytes(br#"<tag  attr-1="/>"  attr2  =  '/>'  3attr/"#)
                         );
                         assert_eq!(position, 42);
                     }
@@ -1983,13 +1986,13 @@ mod test {
                     #[$test]
                     $($async)? fn empty_tag() {
                         let buf = $buf;
-                        let mut position = 1;
-                        let mut input = b"/ >".as_ref();
-                        //                   ^= 4
+                        let mut position = 0;
+                        let mut input = &b"</ >"[$skip..];
+                        //                     ^= 4
 
                         assert_eq!(
                             Bytes($source(&mut input).read_with(ElementParser::default(), buf, &mut position) $(.$await)? .unwrap()),
-                            Bytes(b"/ ")
+                            Bytes(b"</ ")
                         );
                         assert_eq!(position, 4);
                     }
@@ -1997,13 +2000,13 @@ mod test {
                     #[$test]
                     $($async)? fn normal() {
                         let buf = $buf;
-                        let mut position = 1;
-                        let mut input = b"/tag>".as_ref();
-                        //                     ^= 6
+                        let mut position = 0;
+                        let mut input = &b"</tag>"[$skip..];
+                        //                       ^= 6
 
                         assert_eq!(
                             Bytes($source(&mut input).read_with(ElementParser::default(), buf, &mut position) $(.$await)? .unwrap()),
-                            Bytes(b"/tag")
+                            Bytes(b"</tag")
                         );
                         assert_eq!(position, 6);
                     }
@@ -2011,13 +2014,13 @@ mod test {
                     #[$test]
                     $($async)? fn empty_ns_empty_tag() {
                         let buf = $buf;
-                        let mut position = 1;
-                        let mut input = b"/:>".as_ref();
-                        //                   ^= 4
+                        let mut position = 0;
+                        let mut input = &b"</:>"[$skip..];
+                        //                     ^= 4
 
                         assert_eq!(
                             Bytes($source(&mut input).read_with(ElementParser::default(), buf, &mut position) $(.$await)? .unwrap()),
-                            Bytes(b"/:")
+                            Bytes(b"</:")
                         );
                         assert_eq!(position, 4);
                     }
@@ -2025,13 +2028,13 @@ mod test {
                     #[$test]
                     $($async)? fn empty_ns() {
                         let buf = $buf;
-                        let mut position = 1;
-                        let mut input = b"/:tag>".as_ref();
-                        //                      ^= 7
+                        let mut position = 0;
+                        let mut input = &b"</:tag>"[$skip..];
+                        //                        ^= 7
 
                         assert_eq!(
                             Bytes($source(&mut input).read_with(ElementParser::default(), buf, &mut position) $(.$await)? .unwrap()),
-                            Bytes(b"/:tag")
+                            Bytes(b"</:tag")
                         );
                         assert_eq!(position, 7);
                     }
@@ -2039,13 +2042,13 @@ mod test {
                     #[$test]
                     $($async)? fn with_attributes() {
                         let buf = $buf;
-                        let mut position = 1;
-                        let mut input = br#"/tag  attr-1=">"  attr2  =  '>'  3attr>"#.as_ref();
-                        //                                                         ^= 40
+                        let mut position = 0;
+                        let mut input = &br#"</tag  attr-1=">"  attr2  =  '>'  3attr>"#[$skip..];
+                        //                                                           ^= 40
 
                         assert_eq!(
                             Bytes($source(&mut input).read_with(ElementParser::default(), buf, &mut position) $(.$await)? .unwrap()),
-                            Bytes(br#"/tag  attr-1=">"  attr2  =  '>'  3attr"#)
+                            Bytes(br#"</tag  attr-1=">"  attr2  =  '>'  3attr"#)
                         );
                         assert_eq!(position, 40);
                     }