+++ src/reader/slice_reader.rs
@@ -408,6 +408,7 @@ mod test {
         read_event_impl,
         read_until_close,
         identity,
+        0,
         ()
     );
 }